+++ packages/desktop/views/dashboard/index.js
@@ -1,7 +1,7 @@
 export * from './collectibles'
 export * from './developer'
 export * from './settings'
-export * from './wallet'
+export * from './old-wallet'
 export * from './governance'
 export * from './send-flow'
