--- packages/desktop/views/dashboard/new-dashboard/index.ts
@@ -1,1 +1,0 @@
-export { default as NewDashboard } from './NewDashboard.svelte'
