+++ packages/desktop/views/dashboard/old-wallet/index.js
@@ -1,0 +1,1 @@
+export { default as OldWallet } from './OldWallet.svelte'
