+++ packages/desktop/views/dashboard/wallet/Wallet.svelte
@@ -1,42 +1,27 @@
 <script lang="ts">
-    import { Pane, ReceiveAddressButton, TokenList } from '@ui'
-    import { AccountActivity, AccountSummary, SendButton } from '@components'
-    import { selectedAccountTokens } from '@core/token/stores'
+    import { Pane } from '@ui'
+    import { AccountSummaryPane } from './panes'
     import { selectedAccount } from '@core/account/stores'
-    import features from '@features/features'
+    import TabSection from './tab-section/TabSection.svelte'
 </script>
 
 {#if $selectedAccount}
-    <wallet-container
-        class="w-full h-full flex flex-nowrap p-8 relative flex-1
-        bg-gray-50 dark:bg-gray-900 justify-center items-center"
-    >
+    <wallet-container>
         {#key $selectedAccount?.index}
-            <div class="h-full grid grid-cols-3 gap-x-4 min-h-0 min-w-0 max-w-7xl">
-                <div class="flex flex-col space-y-4">
-                    <Pane overflow="visible" classes="flex-none">
-                        {#if features?.wallet?.accountSummary?.enabled}
-                            <AccountSummary />
-                        {/if}
-                    </Pane>
-                    <Pane classes="flex flex-col p-6 space-y-6">
-                        {#if features?.wallet?.sendAndReceive?.enabled}
-                            <SendButton />
-                            <ReceiveAddressButton />
-                        {/if}
-                    </Pane>
-                </div>
-                <Pane classes="h-full">
-                    {#if features?.wallet?.assets?.enabled}
-                        <TokenList accountTokens={$selectedAccountTokens} />
-                    {/if}
-                </Pane>
-                <Pane classes="h-full">
-                    {#if features?.wallet?.activityHistory?.enabled}
-                        <AccountActivity />
-                    {/if}
-                </Pane>
-            </div>
+            <AccountSummaryPane account={$selectedAccount} />
+            <Pane classes="flex flex-col flex-grow border border-solid border-gray-100 rounded-b-none">
+                <TabSection />
+            </Pane>
         {/key}
     </wallet-container>
 {/if}
+
+<style lang="scss">
+    wallet-container {
+        height: calc(100vh - var(--navbar-height, 0px) - var(--windows-navbar-height, 0px));
+        @apply p-8 pb-0;
+        @apply flex flex-col gap-8;
+        @apply relative;
+        @apply bg-white dark:bg-gray-900;
+    }
+</style>
