+++ packages/desktop/views/dashboard/wallet/index.ts
