+++ packages/desktop/views/dashboard/wallet/panes/account-summary/AccountSummaryPane.svelte
