+++ packages/desktop/views/dashboard/wallet/panes/account-summary/components/AccountStardustNetworkSummary.svelte
