+++ packages/desktop/views/dashboard/wallet/panes/account-summary/index.ts
