+++ packages/desktop/views/dashboard/wallet/panes/account-summary/interfaces/account-network-summary-props.interface.ts
