+++ packages/desktop/views/dashboard/wallet/panes/account-summary/interfaces/index.ts
