+++ packages/desktop/views/dashboard/wallet/panes/index.ts
