+++ packages/desktop/views/dashboard/wallet/tab-section/TabSection.svelte
