+++ packages/desktop/views/dashboard/wallet/tab-section/activity/ActivityTab.svelte
