+++ packages/desktop/views/dashboard/wallet/tab-section/activity/components/AssetInfo.svelte
