+++ packages/desktop/views/dashboard/wallet/tab-section/activity/components/row-sections/ActivityStatusSection.svelte
