+++ packages/desktop/views/dashboard/wallet/tab-section/index.ts
