+++ packages/desktop/views/dashboard/wallet/tab-section/portfolio/PortfolioTab.svelte
