+++ packages/desktop/views/dashboard/wallet/tab-section/portfolio/components/TokenListRow.svelte
